more sensible re-structure. fix tests
